+++ src/kushi/core.clj
@@ -14,6 +14,7 @@
    [kushi.specs :as specs]
    [kushi.state :as state]
    [kushi.stylesheet :as stylesheet]
+   [kushi.typography :refer [system-font-stacks]]
    [kushi.utils :as util]))
 
 (def KUSHIDEBUG (atom true))
@@ -21,9 +22,17 @@
 (defn kushi-debug
   {:shadow.build/stage :compile-prepare}
   [build-state]
+
+  (util/pprint+ "kushi-debug:garden-vecs-state" @state/garden-vecs-state)
+  (util/pprint+ "kushi-debug:atomic-user-classes" @state/kushi-atomic-user-classes)
+  (util/pprint+ "kushi-debug:atomic-user-classes" (:minu @state/kushi-atomic-user-classes ))
+  (util/pprint+ "kushi-debug:atomic-declarative-classes-used" @state/atomic-declarative-classes-used)
+  (util/pprint+ "kushi-debug:state/user-defined-keyframes" @state/user-defined-keyframes)
+  (util/pprint+ "kushi-debug:state/user-defined-font-faces" @state/user-defined-font-faces)
+
   (let [mode (:shadow.build/mode build-state)]
     #_(when mode
-      (println "(:shadow.build/mode build-state) =>" mode))
+        (println "(:shadow.build/mode build-state) =>" mode))
     (when (not= mode :dev)
       (reset! KUSHIDEBUG false)))
   build-state)
@@ -32,9 +41,9 @@
 (defn style-map->vecs
   [m]
   (when (and (:map-mode? user-config) (map? m))
-   (let [->coll  #(if (coll? %) % [%])
-         classes (some->> m :kushi/class ->coll (map #(->> % name (str ".") keyword)))]
-     (into [] (concat classes (into [] (dissoc m :kushi/class)))))))
+    (let [->coll  #(if (coll? %) % [%])
+          classes (some->> m :. ->coll (map #(->> % name (str ".") keyword)))]
+      (into [] (concat classes (into [] (dissoc m :.)))))))
 
 (defn- scoped-atomic-classname
   "Returns a classname with proper prefixing for scoping.
@@ -77,6 +86,7 @@
 
 (defmacro defclass
   [sym & coll*]
+  (printing/assert-error-if-duplicate-defclass! {:nm sym :form-meta (meta &form)})
   (reset! state/current-macro :defclass)
   (let [defclass-name            (keyword sym)
         coll                     (if (:map-mode? user-config)
@@ -104,11 +114,11 @@
                                   :garden-vecs garden-vecs}]
 
     #_(util/pprint+
-     "defclass"
-     {:invalid-map-args invalid-map-args
-      :invalid-args invalid-args
-      :styles styles
-      :m m})
+       "defclass"
+       {:invalid-map-args invalid-map-args
+        :invalid-args invalid-args
+        :styles styles
+        :m m})
 
     ;; Print any problems to terminal
     (printing/console-warning-defclass console-warning-args)
@@ -130,9 +140,20 @@
               js/console
               (kushi.core/js-warning-defclass ~console-warning-args))))
          nil)
-      `(do nil)
-      )))
+      `(do nil))))
 
+(defmacro clean!
+  "Removes all existing styles that were injected into #_kushi-dev_ style tag at dev time.
+   Intended to be called by the projects main/core ns on every save/reload."
+  []
+  (let [log? (:log-clean!? user-config)]
+    `(when ^boolean js/goog.DEBUG
+       (do (when ~log? (js/console.log "kushi.core/cleaning!"))
+           (let [sheet# (.-sheet (js/document.getElementById "_kushi-dev_"))
+                 rules# (.-rules sheet#)
+                 rules-len# (.-length rules#)]
+             (clojure.core/doseq [idx# (clojure.core/reverse (clojure.core/range rules-len#))]
+               (.deleteRule sheet# idx#)))))))
 
 (defmacro add-font-face
   "Example:
@@ -147,6 +168,29 @@
          (garden/css (at-font-face m))))
 
 
+
+(defmacro add-system-font-stack
+  [& weights*]
+  (let [weights (if (empty? weights*)
+                  system-font-stacks
+                  (reduce (fn [acc v]
+                            (if (contains? system-font-stacks v)
+                              (assoc acc v (get system-font-stacks v))
+                              acc))
+                          {}
+                          weights*))]
+    (doseq [[weight fonts-by-style] weights]
+      (doseq [[style fonts] fonts-by-style]
+        (reset! state/current-macro :add-font-face)
+        (swap! state/user-defined-font-faces
+               conj
+               (garden/css
+                (at-font-face
+                 {:font-family "system"
+                  :font-style (name style)
+                  :font-weight weight
+                  :src (mapv #(str "local(\"" % "\")") fonts)})))))))
+
 (defn- keyframe [[k v]]
   (let [frame-key (if (vector? k)
                     (string/join ", " (map name k))
@@ -158,11 +202,11 @@
                    v)]
     [frame-key frame-val]))
 
-
-(defmacro defkeyframes [k & frames*]
+(defmacro defkeyframes [nm & frames*]
+  (printing/assert-error-if-duplicate-keyframes! {:nm nm :form-meta (meta &form)})
   (reset! state/current-macro :defkeyframes)
   (let [frames (mapv keyframe frames*)]
-    (swap! state/user-defined-keyframes assoc k frames)))
+    (swap! state/user-defined-keyframes assoc (keyword nm) frames)))
 
 (defn cssfn [& args]
   (cons 'cssfn (list args)))
@@ -193,6 +237,8 @@
       {:styles+classes (drop-last args) :attr (last args)}
       {:styles+classes args :attr nil})))
 
+(def meta-ks [:ancestor :prefix :ident :element :data-attr-name])
+
 (defn- parse-attr+meta [args]
   (let [{styles+classes*
          :styles+classes
@@ -200,12 +246,9 @@
          :attr
          invalid-map-args
          :invalid-map-args}        (style&classes+attr args)
-        meta-ks                    [:parent :prefix :f :fn :ident :element :data-ns]
-        {f            :f
-         component-fn :fn
-         ident        :ident
+        {ident        :ident
          :as          meta}        (select-keys attr* meta-ks)
-        data-ns-key                (or (:data-ns-key user-config) :data-ns)
+        data-attr-name             (or (:data-attr-name user-config) :data-cljs)
         attr                       (apply dissoc attr* meta-ks)
         styles+classes             (if (:map-mode? user-config) (style-map->vecs styles+classes*) styles+classes*)
         {:keys [valid invalid]}    (util/reduce-by-pred #(s/valid? ::specs/kushi-arg %) styles+classes)
@@ -216,28 +259,28 @@
         styles                     (into [] (concat styles* classes-with-mods-hydrated))
         invalid-args               (into [] (concat invalid-map-args invalid))]
 
-    #_(util/pprint+ "parse-attr+meta"
-            {:attr*          attr*
-             :attr           attr
-             :meta           meta
-             :styles+classes* styles+classes*
-             :styles+classes styles+classes
-             :styles*        styles
-             :classes*       classes*
-             :f              (or f component-fn)
-             :ident          ident
-             :data-ns-key    data-ns-key
-             :invalid-map-args   invalid-map-args
-             :invalid-args   invalid-args})
+    #_(util/pprint+
+       "parse-attr+meta"
+       {:attr*          attr*
+        :attr           attr
+        :meta           meta
+        :styles+classes* styles+classes*
+        :styles+classes styles+classes
+        :styles*        styles
+        :classes*       classes*
+        :f              (or f component-fn)
+        :ident          ident
+        :data-attr-name data-attr-name
+        :invalid-map-args   invalid-map-args
+        :invalid-args   invalid-args})
 
     {:attr           attr
      :meta           meta
      :styles+classes styles+classes
      :styles*        styles
      :classes*       classes*
-     :f              (or f component-fn)
      :ident          ident
-     :data-ns-key    data-ns-key
+     :data-attr-name data-attr-name
      :invalid-args   invalid-args}))
 
 (defn classlist [meta classes* selector*]
@@ -263,8 +306,7 @@
                 attr
                 meta
                 ident
-                f
-                data-ns-key]}        (parse-attr+meta args)
+                data-attr-name]}     (parse-attr+meta args)
         {:keys [selector selector*]} (selector/selector-name meta)
         classlist-map                (classlist meta classes* selector*)
         styles                       (parse/+vars styles* selector*)
@@ -275,125 +317,148 @@
         attr-base                    (or attr {})]
 
     #_(util/pprint+
-     "sx*"
-     {:classlist-map classlist-map
-      :styles*       styles*
-      :styles        styles
-      :css-vars      css-vars
-      :tokenized-styles tokenized-styles
-      :grouped-by-mqs grouped-by-mqs
-      :garden-vecs   garden-vecs})
+       "sx*"
+       {:selector* selector*
+        :selector  selector
+        :classlist-map classlist-map
+        :styles*       styles*
+        :styles        styles
+        :css-vars      css-vars
+        :tokenized-styles tokenized-styles
+        :grouped-by-mqs grouped-by-mqs
+        :garden-vecs   garden-vecs})
 
     (merge
      classlist-map
-     {:garden-vecs  garden-vecs
-      :attr         attr
-      :attr-base    attr-base
-      :css-vars     css-vars
-      :f            f
-      :ident        ident
-      :invalid-args invalid-args
-      :data-ns-key  data-ns-key
-      :selector     selector})))
+     {:garden-vecs    garden-vecs
+      :attr           attr
+      :attr-base      attr-base
+      :css-vars       css-vars
+      :ident          ident
+      :invalid-args   invalid-args
+      :data-attr-name data-attr-name
+      :selector       selector})))
+
+(defn- only-attr
+  [args]
+  (let [selected                (:select-ns user-config)
+        select-ns-vector-valid? (s/valid? ::specs/select-ns-vector selected)]
+    (when select-ns-vector-valid?
+      (let [current-ns-sym       (symbol (.toString *ns*))
+            current-ns?          (contains? (into #{} selected) current-ns-sym)
+            current-ns-ancestor? (some? (some #(re-find
+                                                (re-pattern (str %))
+                                                (str current-ns-sym))
+                                              selected))]
+        (when-not (or current-ns-ancestor? current-ns?)
+          (let [attr (or (-> args style&classes+attr :attr) {})]
+            (apply dissoc attr meta-ks)))))))
 
 
 (defmacro sx
-  "Receives n-number of args representing css style declarations or classes.
-   Evaluates styles and adds css data to global registery along with auto-generated
-   (or user-defined) selector prefix.  A optional map containing html attributes
-   and/or kushi-specific options can be passed as the last argument.
-   Returns a normalized attribute map which incorporates the prefixed selector(s)
-   into the :class attribute."
   [& args]
-
-  #_(println "(sx" (string/join " " args)  " ...)")
-
   (reset! state/current-macro :sx)
+  (reset! state/current-sx {:form-meta (meta &form) :args args :bad-mods {}})
+  (or
+   (only-attr args)
+   (let [{:keys [atomic-class-keys
+                 garden-vecs
+                 attr
+                 attr-base
+                 classlist
+                 conditional-class-sexprs
+                 css-vars
+                 ident
+                 invalid-args
+                 data-attr-name
+                 selector]
+          :as   m}               (sx* args)
+         _                       (printing/assert-error-if-duplicate-ident!
+                                  (assoc m :form-meta (meta &form)))
+         styles-argument-display (apply vector args)
+         compilation-warnings    (mapv (fn [v] v) @state/compilation-warnings)
+         invalid-warning-args    {:invalid-args            invalid-args
+                                  :styles-argument-display styles-argument-display
+                                  :form-meta               (meta &form)
+                                  :fname                   "sx"}
+         css-injection-dev       (stylesheet/garden-vecs-injection garden-vecs)
+         og-cls                  (:class attr)
+         cls                     (when og-cls (if (coll? og-cls) og-cls [og-cls]))
+         data-cljs               (let [{:keys [file line column]} (meta &form)]
+                                   (str file ":"  line ":" column))
+         js-args-warning         (printing/preformat-js-warning invalid-warning-args)]
 
-  (let [{:keys [atomic-class-keys
-                garden-vecs
-                attr
-                attr-base
-                classlist
-                conditional-class-sexprs
-                css-vars
-                f
-                ident
-                invalid-args
-                data-ns-key
-                selector]
-         :as   m}               (sx* args)
-        styles-argument-display (apply vector args)
-        compilation-warnings    (mapv (fn [v] v) @state/compilation-warnings)
-        invalid-warning-args    {:invalid-args            invalid-args
-                                 :styles-argument-display styles-argument-display}
-        css-injection-dev       (stylesheet/garden-vecs-injection garden-vecs)
-        og-cls                  (:class attr)
-        cls                     (when og-cls (if (coll? og-cls) og-cls [og-cls]))]
 
     ;; Add classes to previously-used registry
-    (doseq [kw atomic-class-keys]
-      (when-not (contains? @state/atomic-declarative-classes-used kw)
-        (swap! state/atomic-declarative-classes-used conj kw)
-        (printing/diagnostics
-         :defclass-register
-         {:defclass-registered? (contains? @state/atomic-declarative-classes-used kw)})))
+     (doseq [kw atomic-class-keys]
+       (when-not (contains? @state/atomic-declarative-classes-used kw)
+         (swap! state/atomic-declarative-classes-used conj kw)
+         (printing/diagnostics
+          :defclass-register
+          {:defclass-registered? (contains? @state/atomic-declarative-classes-used kw)})))
 
-    (printing/diagnostics
-     :sx
-     (let [style         (:style attr)]
-       {:ident             ident
-        :garden-vecs       garden-vecs
-        :css-injection-dev css-injection-dev
-        :args              args
-        :style-is-var?     (symbol? style)
-        :attr-map          (merge attr-base
-                                  {:class (distinct (concat cls classlist conditional-class-sexprs))
-                                   :style (merge (when (map? style) style) css-vars)})}))
+     (printing/diagnostics
+      :sx
+      (let [style*    (:style attr)
+            style     (merge (when (map? style*) style*) css-vars)
+            class     (distinct (concat cls classlist conditional-class-sexprs))
+            data-cljs {(or data-attr-name :data-cljs) data-cljs}]
+        {:ident             ident
+         :garden-vecs       garden-vecs
+         :css-injection-dev css-injection-dev
+         :args              args
+         :style-is-var?     (symbol? style)
+         :attr-map          (merge attr-base {:class class :style style} data-cljs)}))
 
     ;; Add vecs into garden state
-    (state/add-styles! garden-vecs)
+     (state/add-styles! garden-vecs)
 
-    (printing/console-warning-sx invalid-warning-args)
+     (printing/console-warning-sx invalid-warning-args)
+     (printing/console-warning-sx-mods)
 
-    (reset! state/compilation-warnings [])
+     (reset! state/compilation-warnings [])
 
-    (if @KUSHIDEBUG
-      `(let [og-cls#   (:class ~attr)
-             cls#      (when og-cls# (if (coll? og-cls#) og-cls# [og-cls#]))
-             attr-map# (merge ~attr-base
-                              {:class (distinct (concat cls# (quote ~classlist) ~conditional-class-sexprs))
-                               :style (merge (:style ~attr) ~css-vars)})]
-         (do
-           (when-not (empty? ~compilation-warnings)
-             (.apply
-              js/console.warn
-              js/console
-              (kushi.core/console-warning-number ~compilation-warnings)))
+     (if @KUSHIDEBUG
 
-           (when (seq ~invalid-args)
-             (do
-               (.apply
-                js/console.warn
-                js/console
-                (kushi.core/js-warning-sx ~invalid-warning-args))))
+       ;; dev builds
+       `(let [og-cls#   (:class ~attr)
+              cls#      (when og-cls# (if (coll? og-cls#) og-cls# [og-cls#]))
+              attr-map# (merge ~attr-base
+                               {:class (distinct (concat cls# (quote ~classlist) ~conditional-class-sexprs))
+                                :style (merge (:style ~attr) ~css-vars)
+                                :data-cljs ~data-cljs})
+              logfn#    (fn [f# js-array#] (.apply js/console.warn js/console (f# js-array#)))]
+          (do
+            (when-not (empty? ~compilation-warnings)
+              (.apply
+               js/console.warn
+               js/console
+               (kushi.core/console-warning-number ~compilation-warnings)))
 
-           (kushi.core/inject-style-rules (quote ~css-injection-dev) ~selector)
-           (merge (when ~f {~data-ns-key (kushi.core/ns+ ~f ~ident)})
-                  attr-map#)))
-      (let
-       [{:keys [only-class? only-class+style?]} (util/analyze-attr m)]
-        (cond
-          only-class?
-          `(do {:class (quote ~classlist)})
+            (when (seq ~invalid-args) (logfn# cljs.core/to-array ~js-args-warning))
+            #_(when (seq ~invalid-args)
+                (.apply
+                 js/console.warn
+                 js/console
+                 (cljs.core/to-array ~js-args-warning)))
 
-          only-class+style?
-          `(do {:class (quote ~classlist)
-                :style (merge (:style ~attr) ~css-vars)})
-          :else
-          `(let [og-cls#   (:class ~attr)
-                 cls#      (when og-cls# (if (coll? og-cls#) og-cls# [og-cls#]))
-                 attr-map# (merge ~attr-base
-                                  {:class (distinct (concat cls# (quote ~classlist) ~conditional-class-sexprs))
-                                   :style (merge (:style ~attr) ~css-vars)})]
-             attr-map#))))))
+            (kushi.core/inject-style-rules (quote ~css-injection-dev) ~selector)
+            attr-map#))
+
+      ;; release builds
+       (let
+        [{:keys [only-class? only-class+style?]} (util/analyze-attr m)]
+         (cond
+           only-class?
+           `(do {:class (quote ~classlist)})
+
+           only-class+style?
+           `(do {:class (quote ~classlist)
+                 :style (merge (:style ~attr) ~css-vars)})
+           :else
+           `(let [og-cls#   (:class ~attr)
+                  cls#      (when og-cls# (if (coll? og-cls#) og-cls# [og-cls#]))
+                  attr-map# (merge ~attr-base
+                                   {:class (distinct (concat cls# (quote ~classlist) ~conditional-class-sexprs))
+                                    :style (merge (:style ~attr) ~css-vars)})]
+              attr-map#)))))))