+++ src/kushi/core.cljs
@@ -1,19 +1,7 @@
 (ns ^:dev/always kushi.core
-  (:require-macros [kushi.core])
-  (:require
-   [clojure.string :as string]))
+  (:require-macros [kushi.core :refer [sx clean! add-font-face defclass defkeyframes add-system-font-stack]])
+  (:require [clojure.string :as string]))
 
-(defn clean!
-  "Removes all existing styles that were injected into #_kushi-dev_ style tag at dev time.
-   Intended to be called by the projects main/core ns on every save/reload."
-  []
-  (when ^boolean js/goog.DEBUG
-   (do (js/console.log "kushi.core/clean!")
-       (let [sheet (.-sheet (js/document.getElementById "_kushi-dev_"))
-             rules (.-rules sheet)
-             rules-len (.-length rules)]
-         (doseq [idx (reverse (range rules-len))]
-           (.deleteRule sheet idx))))))
 
 (defn cssfn? [x]
   (and (list? x)
@@ -173,10 +161,6 @@
       (interleave (repeat (/ (- number-of-formats 4) 2) "color:black;font-weight:bold")
                   (repeat (/ (- number-of-formats 4) 2) "color:default;font-weight:normal"))))))
 
-(defn js-warning-sx [m*]
-  (let [m (assoc m* :fname "sx")]
-    (js-warning* m)))
-
 (defn js-warning-defclass [m*]
   (let [m (assoc m* :fname "defclass")]
     (js-warning* m)))
@@ -203,48 +187,7 @@
      (concat
       [warning]
       (interleave (repeat (/ number-of-formats 2) "color:black;font-weight:bold")
-                  (repeat (/ number-of-formats 2) "color:default;font-weight:normal")))))
-                    ;;
-  )
-
-(defn ns+
-  "Creates a string that represents a fully namespaced-qualified
-   identifier for an element within a component rendering function.
-   String includes line number of parent function. This string is
-   used as the value of a custom-data attribute, in order to help
-   quickly identify the associated namespace of the specific element
-   when inspecting output in an environment such as Chrome DevTools.
-
-   Intended to be called with 1 or 2 arguments:
-   1) A var-quoted name of the enclosing component rendering function.
-   2) (Optional) A user-defined keyword, which is a semantic name associated with the html element.
-
-
-   Example:
-
-    (defn my-button [label]
-      [:div
-      (s+ {:cursor :pointer
-           :text-align :center
-           :border [[1 :solid :blue]}
-          {:role :button
-           :on-click #()
-           :data-ns (ns+ #'my-button :outer)})
-        [:span
-         (s+ {:background :yellow}
-             {:data-ns (ns+ #'my-button :inner})
-         label]])"
-
-  ([x]
-   (when (= cljs.core/Var (type x))
-     (ns+ x nil)))
-  ([var-quoted-fn el-ident]
-   (let [{ns* :ns name* :name line* :line} (meta var-quoted-fn)
-         namespace* (when ns* (str ns* "/"))
-         fn-name (when name* (str name*))
-         el-ident-str (when el-ident (str (when fn-name "::") (name el-ident)))
-         line-number (when line* (str ":" line*))]
-     (str namespace* fn-name el-ident-str line-number))))
+                  (repeat (/ number-of-formats 2) "color:default;font-weight:normal"))))))
 
 (defn- merge-with-style-warning
   [v k n]